+++ packages/terra-list/src/MultiSelectList.jsx
@@ -13,6 +13,10 @@ const propTypes = {
    */
   isDivided: PropTypes.bool,
   /**
+   * Whether or not the child items should be displayed in sections with header and list of elements.
+   */
+  hasSections: PropTypes.bool,
+  /**
    * A callback event that will be triggered when selection state changes.
    */
   onChange: PropTypes.func,
@@ -25,6 +29,7 @@ const propTypes = {
 const defaultProps = {
   children: [],
   isDivided: false,
+  hasSections: false,
   onChange: undefined,
   maxSelectionCount: undefined,
 };
@@ -33,12 +38,16 @@ class MultiSelectList extends React.Component {
   constructor(props) {
     super(props);
     this.handleOnChange = this.handleOnChange.bind(this);
-    this.state = { selectedIndexes: SelectableList.Utils.initialMultiSelectedIndexes(props.children, props.maxSelectionCount) };
+    if (!this.props.hasSections) {
+      this.state = { selectedIndexes: SelectableList.Utils.initialMultiSelectedIndexes(props.children, props.maxSelectionCount) };
+    } else {
+      this.state = { selectedIndexes: SelectableList.Utils.initialMultiSelectedIndexesWithSections(props.children, props.maxSelectionCount) };
+    }
   }
 
   handleOnChange(event, index) {
     const { children, maxSelectionCount } = this.props;
-    if (SelectableList.Utils.shouldHandleMultiSelect(children, maxSelectionCount, this.state.selectedIndexes, index)) {
+    if (SelectableList.Utils.shouldHandleMultiSelect(children, maxSelectionCount, this.state.selectedIndexes, index, this.props.hasSections)) {
       event.preventDefault();
       const newIndexes = SelectableList.Utils.updatedMultiSelectedIndexes(this.state.selectedIndexes, index);
 
@@ -50,16 +59,21 @@ class MultiSelectList extends React.Component {
   }
 
   render() {
-    const {
-      children, isDivided, onChange, maxSelectionCount, ...customProps
-    } = this.props;
+    const { children, isDivided, onChange, maxSelectionCount, hasSections, ...customProps } = this.props;
+    let maxCount = 0;
+    if (!this.props.hasSections) {
+      maxCount = SelectableList.Utils.validatedMaxCount(children, maxSelectionCount);
+    } else {
+      maxCount = SelectableList.Utils.validatedMaxCountForSections(children, maxSelectionCount);
+    }
     return (
       <SelectableList
         {...customProps}
         isDivided={isDivided}
         onChange={this.handleOnChange}
         selectedIndexes={this.state.selectedIndexes}
-        disableUnselectedItems={this.state.selectedIndexes.length >= SelectableList.Utils.validatedMaxCount(children, maxSelectionCount)}
+        hasSections={hasSections}
+        disableUnselectedItems={this.state.selectedIndexes.length >= maxCount}
       >
         {children}
       </SelectableList>
@@ -70,5 +84,6 @@ class MultiSelectList extends React.Component {
 MultiSelectList.propTypes = propTypes;
 MultiSelectList.defaultProps = defaultProps;
 MultiSelectList.Item = SelectableList.Item;
+MultiSelectList.Section = SelectableList.Section;
 
 export default MultiSelectList;