+++ packages/terra-date-picker/src/DateInput.jsx
@@ -3,9 +3,9 @@ import PropTypes from 'prop-types';
 import 'terra-base/lib/baseStyles';
 import Button from 'terra-button';
 import IconCalendar from 'terra-icon/lib/icon/IconCalendar';
-import Input from 'terra-form/lib/Input';
+import Input from 'terra-form-input';
 import DateUtil from './DateUtil';
-import styles from './DatePicker.scss';
+import styles from './DatePicker.module.scss';
 
 const Icon = <IconCalendar />;